domains: refactor domain creation logic
Application needs to have some interface for domain creation.
Previously it was single call create_domains() that created only
Domain-D by calling bunch of function in board files which is not very
convinient. Also it was hard to extend it with one more domain
(Domain-U), that will be introduced soon.

Now create_domains() is located in single file and call zephyr-xenlib
directly as the logic is same for all boards. It will use dedicated
domain config included on build time.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>