domains: introduce Domain-U creation
Add domain configuration and creation logic for Domain-U in
AoS application. It is controlled by previously introduced option
CONFIG_AOS_ENABLE_DOMU and supports backend configuration for network
and pvblock, which are currently assigned to domain. Domain-U uses
partition on UFS as rootfs, so it is needed to create proper image
during build.

Signed-off-by: Dmytro Firsov <dmytro_firsov@epam.com>